netns: remove conditionals for service settings
Going without the conditionals (like in secure-node.nix) adds
readability and doesn't reduce evaluation performance (in fact, it
even slightly improves performance due to implementation details
of mkIf).

We need to add a missing 'enabled' check in secure-node.nix, so that
when electrs is disabled no onion service is created that requires its
address.