fixup! use setuidgid instead of runuser
setuidgid doesn't use PAM authentication, which is not needed for our
use case.
this makes it simpler and faster than runuser.

when squashing, use the following commit msg:
When running as root, use setuidgid instead of sudo.
As opposed to sudo or doas, setuidgid is a minimal standalone
binary that needs no external configuration.
Also, it's much faster.